common: Create folder and move code to it for organization
