+++ global/global.ts
@@ -1,4 +1,6 @@
-import { BigDecimal, TypedMapEntry, JSONValue, Value, Entity, TypedMap, Result, Wrapped } from '../index'
+import { BigDecimal } from '../common/numbers'
+import { TypedMapEntry, Entity, TypedMap, Result, Wrapped } from '../common/collections'
+import { JSONValue, Value } from '../common/value'
 import { ethereum } from '../chain/ethereum'
 
 export enum TypeId {
