+++ test/entity.ts
@@ -1,4 +1,5 @@
-import { Entity, Bytes, BigDecimal, BigInt } from 'temp_lib/index'
+import { BigDecimal, BigInt } from 'temp_lib/index'
+import { Entity, Bytes } from 'temp_lib/index'
 
 export function testEntity(): void {
   let entity = new Entity()
